+++ polardbx-20200202/alibabacloud_polardbx20200202/models/_create_open_search_account_request.py
@@ -12,12 +12,20 @@ class CreateOpenSearchAccountRequest(DaraModel):
         dbinstance_name: str = None,
         region_id: str = None,
     ):
+        # The name of the account to be created.
+        # 
         # This parameter is required.
         self.account_name = account_name
+        # The password of the account. The password must be 6 to 32 characters in length and must contain at least three of the following character types: uppercase letters, lowercase letters, digits, and special characters (!@#$&%^*()_+-=).
+        # 
         # This parameter is required.
         self.account_password = account_password
+        # The name of the instance.
+        # 
         # This parameter is required.
         self.dbinstance_name = dbinstance_name
+        # The region in which the instance resides.
+        # 
         # This parameter is required.
         self.region_id = region_id