+++ polardbx-20200202/alibabacloud_polardbx20200202/models/_create_service_account_response_body.py
@@ -90,7 +90,7 @@ class CreateServiceAccountResponseBodyData(DaraModel):
         self.already_exists = already_exists
         # The request ID.
         self.request_id = request_id
-        # The service account type.
+        # The type of the service account.
         self.service_account_type = service_account_type
         # The instance status.
         self.status = status