+++ polardbx-20200202/alibabacloud_polardbx20200202/models/_delete_open_search_account_request.py
@@ -11,10 +11,16 @@ class DeleteOpenSearchAccountRequest(DaraModel):
         dbinstance_name: str = None,
         region_id: str = None,
     ):
+        # The name of the account to delete.
+        # 
         # This parameter is required.
         self.account_name = account_name
+        # The instance name.
+        # 
         # This parameter is required.
         self.dbinstance_name = dbinstance_name
+        # The region ID.
+        # 
         # This parameter is required.
         self.region_id = region_id