+++ polardbx-20200202/alibabacloud_polardbx20200202/models/_delete_open_search_whitelist_group_request.py
@@ -11,10 +11,16 @@ class DeleteOpenSearchWhitelistGroupRequest(DaraModel):
         group_id: str = None,
         region_id: str = None,
     ):
+        # The name of the instance.
+        # 
         # This parameter is required.
         self.dbinstance_name = dbinstance_name
+        # The ID of the service group.
+        # 
         # This parameter is required.
         self.group_id = group_id
+        # The ID of the region where the instance resides. > You can call the [DescribeRegions](https://help.aliyun.com/document_detail/196841.html) operation to query the regions supported by PolarDB-X, including region IDs.
+        # 
         # This parameter is required.
         self.region_id = region_id