+++ polardbx-20200202/alibabacloud_polardbx20200202/models/_describe_context_0config_request.py
@@ -10,10 +10,12 @@ class DescribeContext0ConfigRequest(DaraModel):
         dbinstance_name: str = None,
         region_id: str = None,
     ):
-        # 关联的 PolarDB-X 实例名（pxc- 前缀）
+        # The name of the instance.
         # 
         # This parameter is required.
         self.dbinstance_name = dbinstance_name
+        # The region in which the instance resides. > You can call the [DescribeRegions](https://help.aliyun.com/document_detail/196841.html) operation to query the regions supported by PolarDB-X, including region IDs.
+        # 
         # This parameter is required.
         self.region_id = region_id