+++ polardbx-20200202/alibabacloud_polardbx20200202/models/_describe_open_search_instances_request.py
@@ -14,11 +14,18 @@ class DescribeOpenSearchInstancesRequest(DaraModel):
         page_size: int = None,
         region_id: str = None,
     ):
+        # The instance name.
         self.dbinstance_name = dbinstance_name
+        # The maximum number of entries per page for a paging query. Maximum value: 100. Default value: If you do not specify a value or the value is less than 10, the default value is 10. If the value is greater than 100, the default value is 100.
         self.max_results = max_results
+        # The token for the next query.
         self.next_token = next_token
+        # The page number.
         self.page_number = page_number
+        # The page size.
         self.page_size = page_size
+        # The region ID of the instance.
+        # 
         # This parameter is required.
         self.region_id = region_id