+++ polardbx-20200202/alibabacloud_polardbx20200202/models/_restart_open_search_request.py
@@ -10,8 +10,12 @@ class RestartOpenSearchRequest(DaraModel):
         dbinstance_name: str = None,
         region_id: str = None,
     ):
+        # The instance name.
+        # 
         # This parameter is required.
         self.dbinstance_name = dbinstance_name
+        # The region ID.
+        # 
         # This parameter is required.
         self.region_id = region_id